Fix int64, ts partitions. Rework tests
