+++ test/integration/022_bigquery_test/data/incremental_overwrite_date_expected.csv
