+++ test/integration/022_bigquery_test/data/incremental_overwrite_range_expected.csv
@@ -1,0 +1,5 @@
+id,date_int
+10,20200101
+20,20200101
+30,20200102
+40,20200102
