+++ test/integration/022_bigquery_test/data/incremental_overwrite_time_expected.csv
@@ -1,0 +1,5 @@
+id,date_hour
+10,2020-01-01 01:00:00
+20,2020-01-01 01:00:00
+30,2020-01-01 02:00:00
+40,2020-01-01 02:00:00
