+++ test/integration/022_bigquery_test/data/merge_expected.csv
@@ -1,0 +1,7 @@
+id,date_time
+1,2020-01-01 00:00:00
+2,2020-01-01 00:00:00
+3,2020-01-01 00:00:00
+4,2020-01-02 00:00:00
+5,2020-01-02 00:00:00
+6,2020-01-02 00:00:00
